share and add email button in archive list
