+++ components/ToolPanel.js
@@ -1,19 +1,17 @@
 import { useState, Fragment } from 'react';
 import Grid from '@mui/material/Grid';
 import Button from '@mui/material/Button';
-import SendIcon from '@mui/icons-material/Send';
 import DeleteIcon from '@mui/icons-material/Delete';
 import WhatsAppIcon from '@mui/icons-material/WhatsApp';
 import { useUser } from '../hooks/useUser';
-import axios from 'axios';
 import CustomSnackbar from '../components/CustomSnackbar';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { useProfile } from '../lib/firebase';
 import { useSelector } from 'react-redux';
+import SendEmail from './tool/SendEmail';
 
 export default function ToolPanel({ clearText }) {
-  const title = useSelector((state) => state.input.title);
   const defaultText = useSelector((state) => state.input.text);
 
   const [user] = useUser();
@@ -22,21 +20,9 @@ export default function ToolPanel({ clearText }) {
   const uid = user?.uid;
   const profile = useProfile(uid);
   const whatsappPhone = profile?.whatsapp?.find(x => x.active)?.phone;
-  const userEmail = profile?.email?.find(x => x.active)?.to;
 
   const copyText = () => {
     setAlertProps({ ...alertProps, message: 'Copied!' });
-    setEmailAlert(true);
-  };
-
-  const handleSend = () => {
-    const targetEmail = userEmail || user?.email;
-
-    uid && targetEmail && axios.post('/api/email', null,
-      { params: { to: userEmail || user.email, content: defaultText, title } }
-    );
-
-    setAlertProps({ ...alertProps, message: `Email sent to ${targetEmail}!` });
     setEmailAlert(true);
   };
 
@@ -58,9 +44,7 @@ export default function ToolPanel({ clearText }) {
             </Button>
           </Grid>
           <Grid item>
-            <Button size='small' sx={{ boxShadow: 3 }} onClick={() => handleSend()} variant="text" endIcon={<SendIcon />}>
-              {'Send'}
-            </Button>
+            <SendEmail sx={{ boxShadow: 3 }} text={defaultText} showIcon={true}/>
           </Grid>
           <Grid item>
             <Button size='small' sx={{ boxShadow: 3, color: 'green' }} variant="text" endIcon={<WhatsAppIcon />} disabled={!whatsappPhone}>
